Moved login component into compoenents folder and made some changes on the URLs accordingly
